chore: don't run release action on forks
Forks of the main repository can satisfy the preconditions in
auto_release (e.g., when syncing to upstream, if dependencies have been
updated), triggering an attempted release. This likely won't work, since
it requires github and nuget tokens, but it creates a failed workflow
run and may trigger spurious notifications to developers.

This change updates the release workflow to require that it is
executed in the original Chickensoft repository as a precondition,
preventing these failures.